requirements.txt
streamlit==1.15.2
seaborn==0.11.1
plotly
pandas==1.5.2
joblib==0.16.0
xgboost == 1.5.0
scikit-learn==0.23.2
numpy==1.21.5
yfinance==0.2.3
matplotlib==3.5.1
pandas_datareader==0.10.0
keras==2.9.0
tensorflow==2.9.2
plotly_express==0.4.0
newspaper3k


